Pass the `--ci` flag to jest.
Summary:
This is apparently supposed to be used for ci/cd systems.  Well,
jenkins is such a system!  Let's pass it in.  (I think it disables the
auto-creation of snapshot files or something like that, which isn't
something that has been a problem for us but we might as well be
careful about it.)

Issue: https://khanacademy.slack.com/archives/D0AB3GVS9/p1603729829064400

Test Plan: Fingers crossed

Reviewers: john

Reviewed By: john

Subscribers: benkraft, #devops

Differential Revision: https://phabricator.khanacademy.org/D67826