Properly stop the job as soon as tests are done running.
Summary:
When refactoring webapp-test.groovy recently, I tried to simplify
things a lot, especially around the handling of exceptions.  But I got
rid of one (particularly awkward) exception that turns out to still be
necessary: the TestsAreDone "success" exception.  Without that, we
don't optimally handle the situation where when some clients handle
all the tests before the rest of the clients have even started up
their gce instance.  The only way to cancel the gce-startup
immediately is via an failFast plus an exception.  So I add that
exception back in (and catch it since it's not actually a bad
situation, we just want to terminate the `parallel`).

Issue: https://khanacademy.slack.com/archives/C096UP7D0/p1607452369215400

Test Plan:
I re-ran a smoketest with only a few tests, at
   https://jenkins.khanacademy.org/job/deploy/job/e2e-test/19318/flowGraphTable/
I was hoping that the tests would finish before the last test-worker
even started up, but sadly that didn't happen (it's not really in my
control).  But I did get it to terminate the last test-worker before
it had tried to connect to the server, and it got terminated properly
(without failing and without saying "all test-clients failed"), so it
seems like it's working as intended.

Reviewers: #devops, benkraft

Reviewed By: #devops, benkraft

Subscribers: pamela

Differential Revision: https://phabricator.khanacademy.org/D67904