Use a backup technique to finish off clients when the server dies.
We run the test-clients and test-servers in parallel with
failFast=true, so in theory if the server dies jenkins should kill all
the client jobs as well.  But in my experience, that doesn't happen
consistently, see
https://jenkins.khanacademy.org/job/deploy/job/webapp-test/76897/flowGraphTable/

I think that Jenkins does a good job canceling if the parallel job is
in an sh() step, but not always if it's in another built-in step, or
between steps.

Anyway, this is a backup: we mark all the tests as done.  That way the
clients will at worst finish up the tests they're currently running,
and when it comes time to ask for more they will quickly short-circuit
since they see all tests are done.

Issue: https://khanacademy.slack.com/archives/C8Y4Q1E0J/p1607634247460800

Test plan:
I found a test-job where the server error-exited, and "replay"ed it
with this new code:
   https://jenkins.khanacademy.org/job/deploy/job/webapp-test/76922/