Run only one kotlin test at a time
Summary:

Trying to deploy some changes this morning, I kept hitting errors in
webapp-test (see [1],[2],[3]) where the junit results from a test client
weren't made available to the test-server.

I'm not positive what the root cause here, but I did notice this warning
on each of the console logs below as well as indications that the failed
steps in each of the test runs was while a test client was running
multiple kotlin tests.

```
07:42:21  [2021-01-06T12:42:21.337Z] w: Detected multiple Kotlin daemon sessions at build/kotlin/sessions
07:42:21  [2021-01-06T12:42:21.337Z] w: Detected multiple Kotlin daemon sessions at genfiles/kt_build/kotlin/sessions
```

And I also recalled that we have previously had issues with running
multiple kotlin processes in parallel [4], though that was during
deploys rather than tests and so might not be the exact problem here.

This patch tries to fix the issue by only running one kotlin test at a
time, hopefully fixing the problem in the short term.

[1] https://jenkins.khanacademy.org/job/deploy/job/webapp-test/78763/execution/node/1037/log/
[2] https://jenkins.khanacademy.org/job/deploy/job/webapp-test/78762/execution/node/1165/log/
[3] https://jenkins.khanacademy.org/job/deploy/job/webapp-test/78764/execution/node/1094/log/
[4] https://khanacademy.atlassian.net/browse/INFRA-3594

Issue: None

Test plan:
Cross fingers, push to jenkins and see if my branch passes this time!

Auditors: csilvers, #devops