Specify `--prune-tags` every time we specify `--prune` in `git fetch`.
Summary:
We added `--prune` to our jenkins jobs a while ago to make sure the
local refs always match the remote refs (on github).  But when reading
the `git fetch` docs, it seems like `--prune` doesn't actually do that
for tags; you need to specify `--prune-tags` as well.  This commit
does this.

The reason for the second flag is they're worried you have defined
local tags that you never intend to push upstream, and you don't want
pruning to delete those tags.  (That is, you want the behavior "delete
a tag locally if it used to be on the remote but no longer is," but
git doesn't know if a tag used to be on the remote or not.)  However,
we never have local tags we want to keep around in jenkins, so
`--prune-tags` is always safe.

Issue: https://khanacademy.atlassian.net/browse/INFRA-5892

Test Plan: Fingers crossed

Reviewers: #devops, benkraft, ebrown

Reviewed By: #devops, ebrown

Differential Revision: https://phabricator.khanacademy.org/D68832