Get rid of the `-d` flag for rsync.
The idea was we should delete the obsolete videos_xx.json files once
xx is no longer a lite language.  But it's too easy for `-d` to delete
a file we actually do need by mistake, and it's not like these are big
files that we need to be careful to remove when they're obsolete.

Auditors: aasmund

Test plan:
Fingers crossed