FEI-3301: Update weekly-maintenance.sh to not delete ckeditor
Summary:
This will ensure that the ckeditor JS and theme files will remain on the CDN for us to access, so that we won't have to upload them to GCS as part of the deploy process.

Related PRs:
https://github.com/Khan/webapp/pull/479
https://github.com/Khan/ckeditor-dev/pull/8

Issue: FEI-3301

Test Plan: Not sure how to test this, but I assume that this will keep the ckeditor files in place in GCS, per a discussion with @benkraft.

Reviewers: benkraft, csilvers

Reviewed By: benkraft, csilvers

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D69717