Don't test files that only have sync-tag changes.
We already say we don't deploy services if their only change is a
modified sync tag.  Now we say the same for running tests!  A changed
sync-tag can't change how a test runs (it can change linting, but not
testing), so running tests in that case is extraneous.

Test plan:
I manually ran:
   deploy/trivial_diffs.py 4ab70c298f739861700a8272bd3bb4410789f27d 0dadc9e8cb2dbddc455c1117f13a143bd5014e2c > /tmp/triv
   git diff --name-only --diff-filter=ACMRTUB 4ab70c298f739861700a8272bd3bb4410789f27d...0dadc9e8cb2dbddc455c1117f13a143bd5014e2c | fgrep -vx -f /tmp/triv | wc -l
   git diff --name-only --diff-filter=ACMRTUB 4ab70c298f739861700a8272bd3bb4410789f27d...0dadc9e8cb2dbddc455c1117f13a143bd5014e2c | wc -l
and saw that the second had one more line than the first, which is
exactly the length of /tmp/triv!