INFRA-6777 - Update testing jobs to pass if when there are no tets
All of our testing jobs end with a required analysis phase. Unfortunately analysis fails if there are no testing reports to analyze. This can be a good thing, since there are legitimately times where test reports might fail to be generated or saved, which should cause the job to fail. However, there are other times we have entered the testing phase, but have no tests to run - either they've been marked as already run elsewhere in the case of e2e tests, or there are just no tests to run for whatever has changed in the case of unit tests. In these scenarios, there are no test reports generated, and that fails analysis, when in reality there is no problem and the stage should complete correctly.

Here we detect this prior to running the tests, then set a variable to let our analysis phase later know that it can be skipped as it will not find any reports, thus passing the testing job (assuming there was no other failure in the intrim).