Use `xargs -r` to make sure we don't run it if we don't need to.
In my local (ubuntu) testing, `kill -15` with no other args doesn't
give an error.  But it does on jenkins, so add the `-r`.

`xargs -r` isn't supported on os x, but I don't think that's a concern
for jenkins jobs, which are never going to run on a mac.

Test plan:
Will run on jenkins