Notify in the fastly slack channel how to roll back on error.
When a fastly deploy goes wrong, it's a panic situation.  This PR
makes it easier to recover by always emitting, to the `#fastly` slack
channel, the sun command to run to undo the deploy.

Issue: https://github.com/Khan/internal-services/pull/71

Test plan:
Fingers crossed -- will run this on jenkins (with a deploy to fastly-test).