Added go-codecoverage.groovy to /jobs
This job implements code coverage to be reported on via Jenkins + Cobertura per ADR #515.

It will checkout master and run all tests in /services and /pkg and report on their code coverage. This job is for team information purposes and is not used in conjunction with a deploy because it takes a couple hours to run and generate the report. Gaining insight into application-wide coverage should be useful to teams who want to increase their code coverage and quality.

This job has been running on jenkins for some time now. Now that I've worked out the issues, I'll commit it to this project and update the job to reference this groovy script.

Issue: None

Test plan:
- This job has been running on jenkins here: https://jenkins.khanacademy.org/job/misc/job/go-codecoverage/