[AUDIT] Refactor to teardown and setup webapp between steps.
The old script works fine on some hardware, but not on our boxes we run
as the Jenkins job on. Because of this, the box keeps running out of
memory and hanging resulting in the job failing.

The change here isn't particularly pretty, but mostly we just want to
test to see if restarting the webapp solves the memory problem.