Ensure that rollback code is of the right version
During the rollback in our KRD [1] we noticed that we deploy a wrong
version of the dispatch.yaml. We can visibly see the traffic goes
to the new version of the graphql-gateway.

This is due to the local version of the code would still be on the
latest version, and some operations such as updating `dispatch.yaml`
will pick up the config from local code.

This diff ensure that we use
[1] https://jenkins.khanacademy.org/job/deploy/job/deploy-webapp/9990/

Issue: https://khanacademy.slack.com/archives/C8XGW76FQ/p1646079377461529

Test plan:

1. Merge this which should deploy this to jenkins
2. Create a fake deploy that have a harmless route redirection (e.g.
   redirecting a https://khan-academy.appspot.com/trace  to service-testbed)
3. Deploy, set-default and immediately force a rollback.
4. Validate that the dispatch route is not working after the rollback