Make sure the virtualenv comes before /usr/local/bin on $PATH.
My latest change to install pip3 in /usr/local/bin (for keeper) had
the unfortunate effect that it made all the jobs prefer the pip3 in
/usr/local/bin to the python2 pip that lives in the virtualenv.  This
is because we accidentally were putting /usr/local/bin very high up in
the path, above the virtualenv.

I fix this in two ways:
1) I start the virtualenv *after* other path-munging, not before.
2) When we inherit the virtualenv (so we don't need to start it)
   I add the virtualenv dir to the front of PATH again, just in case.

The latter is necessary because in the "already-set-up" case
withVirtualenv leaves PATH alone, but onWorker munges PATH a second
time (so all its stuff is on PATH twice, oh well), which makes it so
the virtualenv isn't first again.  We fight that by prepending the
virtualenv path a second time too.

Issue: https://jenkins.khanacademy.org/job/deploy/job/webapp-test/131666

Test plan:
https://jenkins.khanacademy.org/job/deploy/job/webapp-test/131669,
which has these changes locally (via `replay`), succeeded.