Add login_v3: false to the keeper-config.
We can't deploy right now because keeper is asking for some auth
confirmation.  It turns out that this is because something triggered
it to start using v3 auth instead of v2 auth (probably an accidental
upgrade I did).

Luckily, we can still disable v3 auth via a config option.  The
"right" way to set this option is in the worker image, but this is a
workaround to get things running again while I rebuild the image.

Test plan:
Ran the command manually on a jenkins worker.