Add fake-gcs to the list of services started to build current.sqlite.
We can't run `make serve` in build_current_sqlite.sh because, um, it
didn't work for some reason.  I think INFRA-7279 has more details.  In
any case, this means we need to list the services to start manually.

It's easy to forget to do this, and indeed we did forget when we added
fake-gcs as a service-dependency in #5869.  This change adds it there.

Note that we will have to get mkcert working no the jenkins machines
as well.  That will be a separate diff (in a separate repo).

Issue: https://khanacademy.slack.com/archives/C01120CNCS0/p1654719166232869

Test plan:
Fingers crossed.