[🔥AUDIT🔥] Clear out the ka-lint cache when running tests on jenkins. (#65)
🖍 _This is an audit!_ 🖍

## Summary:
The ka-lint cache is really useful for avoiding unnecessary lints in
dev.  In theory we could save the same work at deploy-time (and other
times) on jenkins.  But we want to be extra sure everything is ok,
when we run on Jenkins, so I want to suppress the cache there.

The easiest way to do that is to just delete the cache at the
beginning of every test-run, which is what I do in this PR.

Issue: https://khanacademy.slack.com/archives/C02NMB1R5/p1661765927139029

## Test plan:
Fingers crossed

Author: csilvers

Auditors: #devops, #infra-platform

Required Reviewers:

Approved By:

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/65