Set up path on master like we do on the workers.
For some resaon, with the Jenkins upgrade Jenkins no longer reads
`.profile` on the main (jenkins-server) machine, matching the behavior
it's had on the workers for a while.  We worked around that in
`onWorker`.  Now we apply the same workaround in `onMaster`.  This
matches a TODO (though doesn't resolve it, since our workaround is
still a hack).

Issue: https://jenkins.khanacademy.org/job/misc/job/update-ownership-data/1213/

Test plan:
I went to
   https://jenkins.khanacademy.org/job/misc/job/update-ownership-data/1213/
and clicked on `replay` and replaced the contents of onMaster.groovy
with the new version in this PR, and ran it with success.