FEI-4766.2: Tag LambdaTest builds per environment / CI platform
Second change to align on how we tag LambdaTest builds.

The idea is to include tags to builds so we can start filtering builds
by env and platform, so we can start collecting better metrics around
possible flaky tests depending on the tag.

These will be the tags we'll be using from now on:

By environment: `prod` | `preprod` | `znd`.
By CI platform: `github` | `jenkins`.

This PR is responsible for doing the Jenkins-related changes to support
these tags.

In the `jenkins-job` context, we will be supporting the following
combinations:

- `preprod,jenkins` (e.g. first smoke tests or http://prod-someid-somehash.khanacademy.org)
- `prod,jenkins` (e.g. second smoke tests or http://khanacademy.org)

Test plan

Used the Jenkins replay feature to test changes to this job.

- First replay runs against a `preprod` build.

https://jenkins.khanacademy.org/job/deploy/job/e2e-test-cypress/3141/console
https://automation.lambdatest.com/build?build=7684085

- Second replay runs against a `prod` build.

https://jenkins.khanacademy.org/job/deploy/job/e2e-test-cypress/3142/console
https://automation.lambdatest.com/build?build=7684301