FEI-4883: Notify Cypress e2ed test failures to Slack
Ported the logic that we use in the default e2e-test job to notify
failures to Slack using the current Python/Selenium framework.

In the Cypress case, we created a script in `webapp`
(`notify-e2e-results.js`) to be able to replicate the way we send Slack
messages about the e2e test results (for both SUCCESS and FAILURE).

This new script also uses a different Slack wrapper for NodeJS that was
created in the `webapp` repo.

**NOTE:** For more info about the logic created for the JS/Cypress
version of this script, you can take a look at this PR:
https://github.com/Khan/webapp/pull/10867

Test Plan:

1. Navigated to https://jenkins.khanacademy.org/job/deploy/job/e2e-test-cypress/.
2. Clicked on any of the last runs (e.g https://jenkins.khanacademy.org/job/deploy/job/e2e-test-cypress/6343/).
3. Clicked on the `Replay` link.
4. In the `Main Script` textarea, I replaced the contents of that with
   the contents of the `e2e-test-cypress.groovy` file in this PR.
5. Clicked on the `Run` button.
6. Verified that the a Slack notification with the Cypress e2e results
   was sent to a particular thread that I used for testing.

https://khanacademy.slack.com/archives/C096UP7D0/p1673287190526359

Also confirmed that the results were sent to the #cypress-logs-deploys
channel and to my Slack handle (via Direct Message).

**NOTE:** After this is merged, all the test results will be also sent
to the `#qa-log` channel, and only the failures will be sent to the
`#dev-support-log` channel.