[fix] FEI-4883.2: E2E Cypress - Tidy up and fix issues
Re-introduces the changes added in #78, but fixes an issue that was
introduced in the `e2e-test.groovy` file.

The issue was that I didn't properly pass the value of
`USE_FIRSTINQUEUE_WORKERS` as a `boolean`, and instead this was being
passed as a `string`.

This caused the smoke tests to fail, as the `e2e-test.groovy` file threw
the following error:

```
java.lang.ClassCastException: hudson.model.StringParameterValue.value expects class java.lang.String but received class java.lang.Boolean
	at org.jenkinsci.plugins.structs.describable.DescribableModel.coerce(DescribableModel.java:492)
	at org.jenkinsci.plugins.structs.describable.DescribableModel.buildArguments(DescribableModel.java:409)
	at org.jenkinsci.plugins.structs.describable.DescribableModel.instantiate(DescribableModel.java:329)
Caused: java.lang.IllegalArgumentException: Could not instantiate {name=USE_FIRSTINQUEUE_WORKERS, value=true} for hudson.model.StringParameterValue
```

Test plan:

Land these changes and attempt a deploy to verify that both the
`e2e-test` and `e2e-test-cypress` jobs work as expected.