Allow jenkins to update index/queue/pubsub.yaml outside deploy_to_gae.py.
Historically, when there have been changes to index.yaml, queue.yaml,
or pubsub.yaml, that's prompted a deploy of the default module (even
if the python code itself hasn't changed!), and deploy_to_gae.py would
update those yaml files in addition to updating the default module.

Now that python is going away, we want to separate those
responsibilities.  The plan is to create three new pseudo-services,
index.yaml, queue.yaml, and pubsub.yaml.  And should_deploy.py can say
we need to deploy to those services.  Jenkins notices this and uploads
the relevant yaml file to GAE.

This is the jenkins part of things: it notices when the new
pseudo-services are being deployed and uploads the yaml files.  Once
this is tested, I will change should_deploy.py to emit these
pseudo-services when appropriate, and change deploy_to_gae.py to no
longer try to upload to them.

Test plan:
I ran `groovy jobs/build-webapp.groovy` with no new failures.  But my
main test will be to do a deploy after this is in, like so:
```
sun: queue csilvers (SERVICES=queue.yaml,pubsub.yaml,index.yaml)
```
and make sure it runs the upload commands with success.