[🔥AUDIT🔥] Fix the comments to be more accurate. (#83)
🖍 _This is an audit!_ 🖍

## Summary:
I claimed we update queue.yaml in build-webapp and not deploy-webapp,
but that's not really true: we do it in both places.  In the first we
add things, in the second we delete.  This updates the comments to be
clearer.

Issue: none

## Test plan:
none

Author: csilvers

Auditors: #devops, #infra-platform

Required Reviewers:

Approved By:

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/83