Add one default worker for priming Cypress e2e tests
We are currently priming 20 workers that are used by the Python/Selenium
e2e tests. We noticed that the Cypress e2e tests are taking ~2 minutes
before a worker becomes available so the job starts running and this is
due to the Python e2e tests taking up all the available workers.

This PR adds one default worker that will be used by the Cypress e2e
tests so that both `e2e-test` and `e2e-test-cypress` jobs can start
around the same time.

Note that Python and Cypress e2e tests are run in parallel, so we need
to prime+run then at the same time.

Test plan:

I've tested all the Jenkins jobs that are affected by this change using
the Replay feature. I checked that these changes didn't introduce any
regression:

- firstinqueue-priming replay: https://jenkins.khanacademy.org/job/deploy/job/firstinqueue-priming/2758/console
- e2e-test replay: https://jenkins.khanacademy.org/job/deploy/job/e2e-test/45411/console
- e2e-test-cypress replay: https://jenkins.khanacademy.org/job/deploy/job/e2e-test-cypress/6899/console
- webapp-test replay: https://jenkins.khanacademy.org/job/deploy/job/webapp-test/169906/console