Added work for creating cloud run tasks in jenkins
In order for Cloud Run services to handle traffic for versions that don't exist in Cloud Run we need to create traffic tags. These traffic tags is how we can send traffic to default revisions.  We need to do this when we build webapp and when we deploy znds.

Issue: https://khanacademy.atlassian.net/browse/INFRA-8963

Test plan:
Im going to kick off a znd build for service-testbed and check the logs.