FEI-4923.2: Include Slack token in the lambdatest step (#89)
## Summary:

We made some changes in https://github.com/Khan/webapp/pull/11664 to be able to
identify if the LambdaTest platform is going to hit queue timeout errors.

This happens when there are multiple builds in parallel, causing our e2e tests
to fail b/c there are no more available machines to run the tests.

This check was introduced in the `webapp` repo, and the only thing that we need
to do in Jenkins to support it is to include the Slack token so we can send a
notification to the FEI team in case we hit this queue timeout error.

Issue: FEI-4923

## Test plan:

Using replay, I verified that these changes worked fine:

https://jenkins.khanacademy.org/job/deploy/job/e2e-test-cypress/7050/console

Author: jandrade

Reviewers: jeresig, #devops

Required Reviewers:

Approved By: jeresig

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/89