Change most of our cron-based jenkins jobs to send to slack on success.
We had a problem where the `clean_queues` job deleted a taskqueue it
shouldn't have, and it took us a while to track down because nobody
left at the company even realized this job existed!

This PR tries to help with that a little by having most cron-based
jenkins jobs send to slack on success, rather than just on failure.
That way at least people will have a chance of knowing the job exists.

The only cron-jobs that I did not modify are `make-allcheck` and
`find-failing-taskqueue-tasks`, since they send to slack on success in
any case (as part of their job output).

Issue: https://khanacademy.atlassian.net/browse/INFRA-9008

Test plan:
Fingers crossed