Disable the clean-queues.groovy cronjob.
We will no longer run it.  Amos already disabled this job via the UI,
but this change will make sure it doesn't come back again (until we
want it to).

The issue is the job deleted some queues that were still in use.  We
are being very careful with the code before turning this back on again
(if we ever do).

Issue: https://khanacademy.atlassian.net/browse/INFRA-9008

Test plan:
None