Add a job to start a vm for the purposes of running dev-appserver.
This just starts up a vm and tells the user to run some commands to
copy their files over to the vm and then run hotel on it.  I included
some timeout features as well.

Test plan:
I ran this at
   https://jenkins.khanacademy.org/job/misc/job/phabricator-test/92
using the "replay" feature to put in my new code insteaad.