FEI-4888.2: Remove Cypress unused jobs
Summary:

Follow up to #98 that is responsible for removing the Cypress jobs that
we created for development and testing purposes. Now that we have
switched/moved everything to the `e2e-test` job, we can remove the
files/jobs contained in this PR.

NOTE: Doing this in a separate PR to avoid any issues with the current
Jenkins jobs.