FEI-4888.1: Switch to Cypress for e2e tests (#98)
## Summary

This is a rollout of the Cypress e2e tests. This is the first
step in the rollout, which is to switch the e2e tests to run on the
Cypress infrastructure instead of the Python-Selenium infra.

The next step will be to remove the `e2e-cypress-*` jobs that were
created while we tested this integration in parallel with the
Python-Selenium infra.

Issue: FEI-4888

## Test Plan

- [x] Run the `e2e-test` job on this branch and verify that it passes in
  Jenkins via the "Replay" button.

NOTE: I created two separate Replay jobs to verify that `first` and
`second` e2e tests pass.

1. first-smoke-test: https://jenkins.khanacademy.org/job/deploy/job/e2e-test/47341/console
2. second-smoke-test: https://jenkins.khanacademy.org/job/deploy/job/e2e-test/47343/console

- [x] Run the `firstinqueue-priming` job on this branch and verify that
  it passes in Jenkins via the "Replay" button.

https://jenkins.khanacademy.org/job/deploy/job/firstinqueue-priming/3019/console

Author: jandrade

Reviewers: jandrade, jeresig, csilvers, somewhatabstract, dbraley

Required Reviewers:

Approved By: jeresig, csilvers, somewhatabstract, dbraley

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/98