Get rid of unused jenkins jobs and scripts.
I found the unused jobs by running this command on jenkins-server, as
the jenkins user:
```
grep -roh 'jobs/[^ ]*.groovy' ~/jobs/*/jobs/*/*.xml | sort -u
```
This shows all the groovy scripts that are run by currently configured
jobs.

I then compared this against what's in github, and removed the three
jobs that were in github but not in a config file.

Then for every file in the top-level jenkins-jobs repo, I ran
`git grep <filename>`, and deleted all the ones with no hits.
That was most of them!

Issue: none

Test plan:
Fingers crossed, we can always revert if we need to