Keep a metric ton of delete-versions logs.
We run delete-versions continuously -- probably because we have so
many versions at the moment, we could really do a better job of
managing that -- and it's not enough to debug.  Let's keep a lot more
logs so we can debug better.  The logs are short so it should be ok.

Issue: https://khanacademy.slack.com/archives/C01120CNCS0/p1683676021846129

Test plan:
None