Introducing succeeded, failed, and aborted to monitoring messaging
Previously monitoring was only discerned as "finished" making it difficult
for buildmaster to determine whether there were monitoring issues or not.
This should now provide some better, still not perfect, idea to buildmaster.

Issue: "none"

Test plan:

Run a deploy that succeeds, fails monitoring, and where the monitoring stage is aborted.
This is to some degree :fingers_crossed: as the only place to test this is in prod