Make the paths set in onWorker and onMaster match .profile.
Sadly, jenkins doesn't exec `.profile` or `.bashrc` when running
jenkins jobs, it just uses the same path the jenkins binary uses.

On the workers, the jenkins binary (actually the `jar slave.jar`
process) runs via ssh, which *does* exec `.profile`, so all jenkins
jobs on workers run with the path that's in `.profile`.

But on the main jenkins server, the jenkins binary (`jar jenkins.jar`)
runs via systemd, which uses its own default path.  In particular,
systemd runs as root (and later does a proces-chown to the jenkins
user), so it does not use jenkins's .profile.

There might be some way to change this in systemd, but that's pretty
black magic.  Instead, we just set the path we want in our
onWorker.groovy and onMaster.groovy scripts, which run at the
"outside" of all our jenkins jobs.

Unfortunately, the paths set in those scripts had gotten out of date
with the path set in the .profile file.  I get them back in sync, and
add a breadcrumb in .profile so hopefully if the path changes again,
we'll update it everywhere we need to.

This lets me remove the explicit path-adding in
update-devserver-static-images.groovy, which was always an awkward
hack.

Issue: none

Test plan:
Fingers crossed.  In theory this shouldn't affect any jobs, but we'll
have to land it and see..