Logging for failed http requests to buildmaster
We are having issues aborting jobs and I have traced it back to a race condition between the http request we send to buildmaster for `prompt-status` for `set-default` and when the message for aborting a job is processed by jenkins.  Its unclear to me what error it is we are getting when the request fails so I am adding logging for it.

Issue: https://khanacademy.atlassian.net/browse/INFRA-9495

Test plan:
groovy vars/*.groovy