[🔥AUDIT🔥] Logging for failed http requests to buildmaster (#139)
🖍 _This is an audit!_ 🖍

## Summary:
We are having issues aborting jobs and I have traced it back to a race condition between the http request we send to buildmaster for `prompt-status` for `set-default` and when the message for aborting a job is processed by jenkins.  Its unclear to me what error it is we are getting when the request fails so I am adding logging for it.

Issue: https://khanacademy.atlassian.net/browse/INFRA-9495

## Test plan:
groovy vars/*.groovy

Author: MiguelCastillo

Auditors: #devops, csilvers

Required Reviewers:

Approved By:

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/139