Reduce how much we depend on secrets.py in jenkins-jobs.
We're working to get rid of secrets.py (at least once the SAT product
goes away, and we no longer have any python2 in prod).  One large
remaining client is jenkins, which uses `withSecrets()` in a lot of
places.  This PR works to normalize those uses.

It does this in 2 ways:

1. Some uses are unnecessary: they either call python code that
   doesn't require secrets, or they don't even call python code at all
   anymore.  I just removed those withSecerts calls.

2. Some uses are just for slack: they either call a python script that
   only uses secrets to talk to slack, or they call alertlib directly,
   to talk to slack.  In those cases, I use a new
   withSecrets.forSlackAlertlib only.  Right now that is just an alias
   for `withSecrets`, but in a follow-up PR I'll make it fetch just
   the one secret (and directly from GSM, not from secrets.py).

We're left with withSecrets being used in only two places.  The first,
when deploying the SAT product, is unavoidable but will go away once
we no longer deploy python2.  The second is for a few secripts that
are obsolete but still in source control.  We should delete the
scripts, but not in this PR.

Issue: none

Test plan:
Fingers crossed: I'll try doing a deploy with this on a friday.