Get rid of send-to-email support, which hasn't worked in a long time.
Looking over the alertlib code, I noticed that the secrets that
alertlib needs to send email -- sendgrid_username and
sendgrid_password -- do not exist in secrets.py, and possibly haven't
for a long time.  This means that sending to email would never work.
Given that, in jenkins, we only use send-to-email mode for two jobs
and then only on failure, and given that those jobs haven't failed in
months and months, it's possible that this feature hasn't worked for a
long time and we never noticed.  (Especially since the email goes to
jenkins-admin, which as far as I know I'm the only one who ever looks
at it.)

So let's just get rid of this feature, and depend to sending to slack
on failure.  That seems to have been working well so far.

I considered trying to fix send-to-email instead, by restoring these
sendgrid secrets to secrets.py.  But I couldn't find the secrets in
either gsm or keeper.  Rather than try to track them down or do other
heroic efforts to revive this feature, I think it's better just to let
it Rest In Peace.

Issue: none

Test plan:
Will run on jenkins.