Get rid of the send-to-bugtracker option in jenkins notify(). (#140)
## Summary:
It was only being used in one place, and as far as I can tell that one
place never worked.  Which is good: the allcheck job hasn't passed
in a year and a half, and we don't need a jira task created for that
fact every night.  The existing slack notifications are more than
sufficient.

So let's get rid of it, not just for cleanliness but
so I can clean up the secrets-requirements of jenkins a little bit.

Issue: none

## Test plan:
Fingers crossed

Author: csilvers

Reviewers: nathanjd, csilvers, dbraley

Required Reviewers:

Approved By: nathanjd, dbraley

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/140