Decrypt both slack and stackdriver secrets, in two places.
There are a couple of scripts we run that send data to stackdriver:
monitoring and set-default.  Let's make sure they have the secrets
they need.

While writing this I discovered a bug in how we were accounting for
how the secrests blocks were being used, meaning that we weren't
deleting secrets-files when we were done with them, like we were
supposed to.  I fixed that up too.

Issue: None

Test plan:
Will try a deploy